Update SpriteFont.cs
Removed inconsistency's with drawstring. The lines removed account for height cropping that spritebatch doesn't account for which would result in, a accumulating error between what is drawn and what measure string would report for the total height of the measured text. 
Im doubtful if it is correct to second guess the line height returned by a font if spritebatch did account for it.