Changes the default to 1.
After reading the docs, it appears that the default value on the actual `EnvironmentMapEffect` side is intended to be 1. Making the writer default to the same value seems like a good choice. I also tweaked the comment because the way it is written, something in the documentation is interpreting that as a numbered list when it is just the end of a sentence.