Add Reference Assembly Packages.
When building Library projects or Content Pipeline
it would be better if we referenced a reference
assembly version of the MonoGame.Framework.dll.

We used to have a Portable library which was generated
before the upgrade to .net 6.

This commit adds two new Packages

MonoGame.Framework.Ref
MonoGame.Framework.Content.Pipeline.Ref

These contain generated reference assemblies from
the `jetbrains.refasmer.clitool` [1]. This tool processes
an assembly, stripping out everything but the API and
adding a `ReferenceAssembly` attribute.

[1] https://github.com/JetBrains/Refasmer/