Upgrade Android to handle obsolete API's.
As the underlying Android platform evolves they tend to obsolete
many API's. We should be keeping on top of these so that the
Android implementation stays current. However we cannot remove
the old APi's completely due to the massive number of devices
which Android supports.

So we need to make use of Runtime checks to make sure we call the
correct "supported" API's on the correct versions of Android.
Previously this had to be done by checking the `SdkInt` version.
With .net 6 and beyond we can make use of the
`OperatingSystem.IsAndroidVersionAtLeast` utility method. This is a
method which the compiler is aware of and it can hide the warnings about
obsolete API's around the code it protects.

Also upgraded other obsolete API's which do not need runtime checks.