Detect GamePad/Keyboard events on Android
Fixes #8463

This was a bug, we assumed that ALL KeyDown and KeyUp
events were coming from a GamePad on Android! They might
not, users can use USB or Bluetooth keyboards. As a result
key presses would be passed to the GamePad not the keyboard.

So lets check the type of the device sending the data and
route it to the correct place. This should allow keyboards
to work.