Run the Basic ContentBuilder Tests on CI (#8436)
We are currently not running ANY tests on CI.
Lets correct this by running at least the content building and
processor tests on CI.

Certain content pipelines tests (Audio & Effects) have
been disabled because they fail or hang on CI at this time.

Current the tests run in x64 dotnet so that we can pass the
TextureCompression tests.
This PR #8456 should fix the texture compression issues so we do not
need x64 dotnet.

The Effect compilation for Linux /Mac needs to happen
in wine. But dotnet 8 does not seem to run under wine on CI at this
time. So that will need to be looked at.