Add Self Hosted Runner for Tests (#8493)
This enables the UI unit tests on some self hosted machines we have available.
There were some issues regarding running UI based tests in NUnitLite
via dotnet test on MacOS. This is because you MUST create NSWindow
instances on the Main Thread. So we had to modify the Test runner to
marshal the UI tests over to the main thread.

This added a new [RunOnUI] attribute, which can be used to flag a
test which MUST run on the UI thread. This works for all platforms.

There were still some slight issues with the test on MacOS if running
ALL the UI tests in on go they would fail. But running them in smaller
batches works. I suspect that there is a bug in one of the tests which
is causing this we need to track it down.

A few tests have been disable, issues need to be opened for these so
that they can be fixed.

Also fixed Short2 and NormalizedByte4 types so they actually pass
the tests.

Fixed the casing on the Artifacts folder so it is consistent across
all platforms (Linux and Mac were using lower case in some areas).

Fixed the Adpcm audio support so it unpacks the data correctly.