Allow a Texture to have multiple outputs. (#8524)
On
[Android](https://developer.android.com/guide/playcore/asset-delivery/texture-compression.)
there is a requirement to support multiple Texture compression formats.
With this one change we will not longer ignore items which have a
different output file. This will allow use to do things like this.

```
/outputDir:bin/$(Platform)
/intermediateDir:obj/$(Platform)
/config:
/profile:Reach
/compress:False
/processorParam:TextureFormat=Compressed
/build:ContentFont.spritefont

/importer:TextureImporter
/processor:TextureProcessor
/processorParam:TextureFormat=Compressed
/build:../../Textures/LogoOnly_64px.png;Textures/LogoOnly_64px

$if platform=Android
/processorParam:TextureFormat=Pvrtc
/build:../../Textures/LogoOnly_64px.png;Textures#tcf_pvrtc/LogoOnly_64px

/processorParam:TextureFormat=Astc
/build:../../Textures/LogoOnly_64px.png;Textures#tcf_astc/LogoOnly_64px

/processorParam:TextureFormat=Dxt
/build:../../Textures/LogoOnly_64px.png;Textures#tcf_s3tc/LogoOnly_64px
```

Note the output files are in different sub directories.

Update the editor to allow users to do this. 
Fix a bug in the editor where if you add a directory , hidden files were
included. (e.g .DS_Store on mac).