Install dotnet tools locally in the Nuget Packages (#8527)
In commit 7a398b01 we added two new tools for compressing textures,
`crunch` and `basisu`.
The problem there is that the management of the
`.config/dotnet-tool.json` file by the users was becoming an issue.
We needed a more automatic way to install the required tooling. 

The problem is using the standard `dotnet tool install` calls requires a
`.config/dotnet-tool.json` file to be present in either the current
directory or a directory that is in the path ABOVE the current one. You
would use the `--create-manifest-if-needed` flag to create the manifest,
but that will still leave the users having to manage and upgrade the
.json file every time we do a release.

So lets get the pipeline to install the tooling itself. 
The `dotnet tool install` command has an additional
argument`--tool-path` this allows us to say where we want the tool
installed. Once that has happened we get a native binary launcher in
`--tool-path` which allows us to launch the app directly without using
the `dotnet` executable. So what this allows us to do is install the
tooling locally in the directory that the content pipeline is installed.
This will usually be the global `.nuget/package` directory.

We need to keep an eye on the `DOTNET_ROOT` environment variable when
installing the tooling, just in case a user (or CI) wants to use a
custom dotnet install.

Fix the calls to `basisu` and `crunch` to handle spaces in paths. Same
for other area's which shell out to `mgcb`.

Also added support for overriding which MGCB to use in the editor by
looking for a `MGCBCommand` environment variable. This should be the
full path to either the `mgcb` exe or the `mgcb.dll` that the user wants
to use. This will allow users to change to a custom content compiler
without having to change the editor.

The one downside is users will no longer be able to run `dotnet mgcb`
directly in their project directory unless they install the tooling
manually.

The Editor also needs to download the `mgcb` tool. We place this in the
`ApplicationData` folder rather than in the local assembly folder
because we cannot guarantee that folder is writable. So in this instance
a location specific to the current user seemed like the best approach.

The long term plan is to bundle all these tools into a single native
library which can be called directly by the content pipeline.