Making MonoGame AOT/trimming compliant (#8623)
I started addressing #8104 and those changes should make MonoGame mostly
safe in AOT contexts (it fixes all the errors of loading assets and
removes the need to force MonoGame to not be trimmed).

Working on this actually made visible more AOT issues. I fixed the most
trivial IL warnings but some of them persists.
Also, since BRUTE hasn't fully retired yet, I'm trying to write this in
a ```net452``` backward compatible way (some of the newer and better
reflection stuff isn't available on older .NET Frameworks).

Any help is welcome!

Here are my notes on the remaining issues (none really are blockers,
just enhancing the AOT experience and avoid misleading warnings for
users):

- [x] ```ReflectiveReader```, ```DictionaryReader```, ```ArrayReader```,
and ```MultiArrayReader``` need to be flagged as not compatible with AOT
with the ```RequiresDynamicCode``` attribute. Those types are generic
and can't be made safe for AOT. The warning inside the attribute must
encourage users to rewrite their content serialization in a way that
doesn't rely on reflection is they wish to use AOT.

- [ ] WindowsDX will [fail loading an embedded
icon](https://github.com/MonoGame/MonoGame/blob/develop/MonoGame.Framework/Platform/Windows/WinFormsGameWindow.cs#L261)
because ```assembly.Location``` is always null in AOT contexts.

Those ones aren't really issues and will likely not trigger runtime
errors. We could just disable warnings on those, but it would be nice to
try to fix them in a clean way to avoid unforeseen consequences:

- [ ]
[VectorConversion](https://github.com/MonoGame/MonoGame/blob/develop/MonoGame.Framework/Design/VectorConversion.cs)
uses reflection to convert to/from ```IPackedVector```. This needs to be
rewritten reflection-free. We can also assume that ```IPackedVector```
is never trimmed and just disable the warnings, but that's an easy fix
and likely less error-prone to not rely on disabling warnings.

- [ ] There are a couple of instances of type instance creation that we
should try to get rid of if possible:

https://github.com/MonoGame/MonoGame/blob/2e72fcfa4144ea5bc413963af8f97a2d5e902338/MonoGame.Framework/Content/ContentReaders/MultiArrayReader.cs#L36

https://github.com/MonoGame/MonoGame/blob/2e72fcfa4144ea5bc413963af8f97a2d5e902338/MonoGame.Framework/Graphics/Vertices/VertexDeclaration.cs#L214

- [ ] ```ContentTypeReaderManager``` has a warning here (might be
fixable):

https://github.com/MonoGame/MonoGame/blob/2e72fcfa4144ea5bc413963af8f97a2d5e902338/MonoGame.Framework/Content/ContentTypeReaderManager.cs#L154

- [ ] ```ContentManager``` has 2 warnings here (can be disabled, this
will likely never fail, and this is only useful for Android):

https://github.com/MonoGame/MonoGame/blob/2e72fcfa4144ea5bc413963af8f97a2d5e902338/MonoGame.Framework/Content/ContentManager.cs#L592

- [ ] DesktopGL has ```VertexBuffer.GetBufferData()``` which might fail
in AOT depending on the vertex declaration (we can't fix this, but it
would be best to flag this method as incompatible, but like other
warnings, this should actually be fine):

https://github.com/MonoGame/MonoGame/blob/2e72fcfa4144ea5bc413963af8f97a2d5e902338/MonoGame.Framework/Platform/Graphics/Vertices/VertexBuffer.OpenGL.cs#L92