Remove return so that it can try again with wine, instead of wine64. (#8676)
This commit removes an early return in the WineHelper.cs file to allow the process to retry with "wine" as the fallback executable instead of "wine64".

Removed the return statement in the catch block so that after changing the executable to "wine", the program will continue to start the process.

This change enables the intended retry logic for process execution.