Fix .NET 4.5 compatibility and restore ConsoleCheck (#8678)
For some reason, ```ConsoleCheck``` has been unplugged from the CI and
as a result some PRs introduced incompatibilities with .NET 4.5 / C# 5.

This PR fixes the incompatibilities and restores ```ConsoleCheck```.

For now, the ```ConsoleCheck``` only tests against the current code
architecture. It doesn't test the newer ```Native``` structure but some
consoles have already been ported to that structure, so we need
something like ```ConsoleCheck.Native```.

Additionnaly, it would be necessary to have the ```DesktopGL``` tests to
run using ```PublishAot``` to make sure that no PR uses runtime
reflection (which we can't test with building ```ConsoleCheck```) to
avoid regressions on the ```PublishAot``` compatibility.

-----

If anyone wonders "why is this limitation still in place? didn't you
change the console runtime?", the answer is: yes and no.

Yes, the consoles have been moved toward using ```PublishAot``` instead
of BRUTE. All but one, because ```PublishAot``` can't be used there.
Which means that there is one single platform currently limiting the
entirety of MonoGame. We're trying to figuring it out, but it's a
process we don't have a power over and involves bureaucracy (or possibly
to rewrite entirely BRUTE). So no ETA on fixing this, but it's expected
to take a very long time.