Refactor to make accelerometer and compass classes available in cross-platform apps. (#8663)
Fixes #8664 


### Description of Change

Moved the SensorData classes up to the top level (still don't like that
the namespaces are `Microsoft.Devices.*`), then refactored the Platform
classes to remove the iOS IFDEF, following more consistent patterns in
the codebase.
Then added the classes to the non supported classes (the throw
`PlatformNotSupported` exceptions if called), but this now allows game
devs to use Accelerometer and Compass classes in a more cross-platform
way.

Since Microsoft has pulled the WinPhone, in MG 4, we should consider
renaming the namespace to `MonoGame.Devices.*`, IMHO.