Support running MGCB on ARM64. This allows developers to use ARM64 laptops, such as Apple-silicon MacBooks running Windows through Parallels.
While MGCB itself is platform neutral, it loads platform-specific DLLs at runtime for things like shader compilation and texture conversion.

Windows on ARM64 is not currently supported as a native host platform (see #8398), but Windows includes an x86/x64 emulator.

To execute MGCB in an emulated x64 process, and be able to load the normal x64 DLLs, it is necessary to launch the x64 build of dotnet. This can be installed side-by-side with the native ARM64 build of dotnet.