Fixed typo in readme (#8759)
<!--
Are you targeting develop? All PRs should target the develop branch
unless otherwise noted.
-->

Fixes #

<!--
Please try to make sure that there is a bug logged for the issue being
fixed if one is not present.
Especially for issues which are complex. 
The bug should describe the problem and how to reproduce it.
-->

### Description of Change

<!-- 
Enter description of the fix in this section.
Please be as descriptive as possible, future contributors will need to
know *why* these changes are being made.
For inspiration review the commit/PR history in the MonoGame repository.
-->

Fixed typo in readme