Add C Header generation to MGFXC (#8767)
The stock effects for the new native runtimes are embedded in the native
MonoGame library.

Since not all platforms have a form of embedded resources (like Windows)
we need to resort to header files to include a binary stream.

This PR adds the ability to generate a C header file with the embedded
binary effect:


![image](https://github.com/user-attachments/assets/1d5784aa-4c67-4b52-8a28-bb54ac7e4eca)

Once merged we will switch to using this technique for all native
backends.