Reorder workflows to match runs (#8793)
# Description of Change

Restructure of Dependencies install to align both build and test. Each
component in its own block.

This switches each dependency into it's own block of management which
makes it easier to manage each section, especially for the build pass,
rather than a single block of scripts.

This aligns what was happening in the Test path already which seemed
more stable.

---------

Co-authored-by: Dominique Louis <savagesoftware@gmail.com>