Some small improvements to WineHelper.cs
There are a few additional environment variables that can be set
to help with MonoGame development on Wine.

We can expand the items in `WINEDLLOVERRIDES` so that it does not
try to start `explorer.exe` when running the content pipeline.
Also that it does not start `services.exe` either.

The `MVK_CONFIG_LOG_LEVEL` environment variable can be set to
`0` to disable the MoltenVK logging, which can be quite verbose
and distracting when running the content pipeline.

Also added an entry in the launch.json file to allow
the debugging of `MGFX` in VSCode.

Added a new error message to the `WineHelper` class if wine is not found.
//TODO update the url to be something that works.