Add SDL mapping for Key 35 (102 key backslash or UK pound key) (#7945)
See #7488 
Adding this key entry as Oem8 due to lack of suitable existing mappings
in the Oem keys and I wasn't sure it was right to necessarily add a new
one for this use case.

The UK keyboard #/~ key maps as key code 35 in SDL due to its internal
conversion to ascii, however there is no OEM Pound virtual key, and in
native UK keyboard mappings this key maps to OEM 7 (which is OEM Quote
in mono's mappings).
By Mapping this key to Oem8 we can allow support for this key in
DesktopGL for 102 key users, however it may need to be checked/verified
with someone that has a US 102 key layout keyboard to ensure the mapping
comes through correctly.

Co-authored-by: Simon (Darkside) Jackson <darkside@zenithmoon.com>