Fix AOT type signatures and error handling (#8850)
Reviewing #8849 made me to investigate why AOT would fail in some
scenarios.

It appeared that some type signatures were wrong for the type cache
(```SpriteFontReader```, ```IndexBufferReader```, and
```ExternalReferenceReader```).

This PR fixes this and also tries to improve error messages when
NativeAOT fails to retrieve types (but unfortunately recent changes in
NativeAOT make these errors hard/impossible to properly catch, so it is
likely that in the future we'll get reports of AOT not working while
it's actually the error that couldn't be caught to display a nice
message with the explanation and how to fix).