Fix MGCB calling ffmpeg with ffprobe arguments (#8856)
### Fixes

Video content processor now calls ffprobe instead of ffmpeg, as
intended.

### Description of Change
MGCB was incorrectly trying to run ffmpeg with arguments meant for
ffprobe (eg. `-show_format`), causing the project to fail to build when
a video file was included in the assets. This change corrects the method
call to use ffprobe, resolving the issue.