Remove the `RestoreContentCompiler` as it was redundant. (#8815)
* Remove the code which installed tools in a custom folder.
* Added a new `_RestoreMGCBTool` target. This target will restore the
tools, but importantly it will put a flag in the `obj` directory to
track that is has done so. This means that the target will not run on
every build.
* Moved the content build from before `Build` to before `Compile` but
after `ResolveProjectReferences` , this allows any dependencies to be
built first. So if you reference a content pipeline extension it will be
built before the content is built.