Fix bounds calculation on blacking out pixels (#8918)
Fixes #8917

### Description of Change
All I've done is to make sure the code uses the byte bounds instead of
the pixel bounds when blacking out the image.

## Issue
Previously, the code would just stop processing after a quarter. An
example with a 3x3 image gives:

w = 3
h = 3

Such that we cover pixel space indices [0,9).

However, we would then increment the index variable by 4 as if we were
in byte space, terminating after 25% of the processing.

## Bonus
I also took the liberty of making the comments more concise.