[CI] Hook up shader compilation to build (#8916)
This is the first part to making shader compilation be done on the CI.

The 2nd part will automatically commit the files so a fresh repo clone
would not require shader recompilation.

~~I've removed the RebuildMGFX.bat scripts, would you like me to make a
build_shaders.bat/.sh on top level of the repo by any chance? ~~

Added scripts folder to the repo root with new versions of RebuildMGFX

~~DX12 shader build step weren't working and I can't easily get to the
bottom of it from macOS / Linux :(~~

This PR is needed to unblock
https://github.com/MonoGame/MonoGame/pull/8903 from getting reviewed and
merged

CC. @tomspilman @ThomasFOG