Make a Better Error Message for Audio Imports (#8932)
We currently get a really odd error message if we hit a problem
processing audio files.

`Ensure the file is a valid audio file and is not DRM protected`

Allot of the time it has nothing to do with DRM. The underlying errors
are ignored.
This commit reworks the error messaging to produce a error message which
includes the actual error.

Also updated an error message that would give some weird error when
using mp3/ogg with the WavImporter. We can do better there.

Fixed an bug in the launch.json so that we get a process list to pick
the process we want to attach to. Very handy for debugging mgcb.